backend/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
google-cloud-translate==3.11.0
google-cloud-texttospeech==2.13.0
requests==2.31.0
python-dotenv==1.0.0
